Move document-dto mapping methods in OpenSearchService.
