+++ server/src/main/kotlin/org/octopusden/octopus/vcsfacade/document/RefDocument.kt
@@ -1,9 +1,6 @@
 package org.octopusden.octopus.vcsfacade.document
 
-import org.octopusden.octopus.vcsfacade.client.common.dto.Branch
-import org.octopusden.octopus.vcsfacade.client.common.dto.Ref
 import org.octopusden.octopus.vcsfacade.client.common.dto.RefType
-import org.octopusden.octopus.vcsfacade.client.common.dto.Tag
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.data.elasticsearch.annotations.Document
 import org.springframework.data.elasticsearch.annotations.Field
@@ -26,14 +23,4 @@ class RefDocument(
 
     override fun toString() =
         "RefDocument(id=$id, repository=$repository, type=$type, name=$name, hash=$hash, link=$link)"
-
-    fun toDto() = when (type) {
-        RefType.BRANCH -> Branch(name, hash, link, repository.toDto())
-        RefType.TAG -> Tag(name, hash, link, repository.toDto())
-    }
-
-    companion object {
-        fun Ref.toDocument(repositoryDocument: RepositoryDocument) =
-            RefDocument(repositoryDocument, type, name, commitId, link)
-    }
 }
