+++ server/src/main/kotlin/org/octopusden/octopus/vcsfacade/service/OpenSearchService.kt
@@ -3,11 +3,19 @@ package org.octopusden.octopus.vcsfacade.service
 import org.octopusden.octopus.vcsfacade.client.common.dto.Branch
 import org.octopusden.octopus.vcsfacade.client.common.dto.Commit
 import org.octopusden.octopus.vcsfacade.client.common.dto.PullRequest
+import org.octopusden.octopus.vcsfacade.client.common.dto.PullRequestReviewer
+import org.octopusden.octopus.vcsfacade.client.common.dto.Ref
+import org.octopusden.octopus.vcsfacade.client.common.dto.RefType
+import org.octopusden.octopus.vcsfacade.client.common.dto.Repository
 import org.octopusden.octopus.vcsfacade.client.common.dto.SearchSummary
+import org.octopusden.octopus.vcsfacade.client.common.dto.Tag
+import org.octopusden.octopus.vcsfacade.client.common.dto.User
 import org.octopusden.octopus.vcsfacade.document.CommitDocument
 import org.octopusden.octopus.vcsfacade.document.PullRequestDocument
+import org.octopusden.octopus.vcsfacade.document.PullRequestReviewerDocument
 import org.octopusden.octopus.vcsfacade.document.RefDocument
 import org.octopusden.octopus.vcsfacade.document.RepositoryDocument
+import org.octopusden.octopus.vcsfacade.document.UserDocument
 import org.octopusden.octopus.vcsfacade.dto.VcsServiceType
 
 
@@ -32,4 +40,63 @@ interface OpenSearchService {
     fun findCommitsByIssueKey(issueKey: String): List<Commit>
     fun findPullRequestsByIssueKey(issueKey: String): List<PullRequest>
     fun findByIssueKey(issueKey: String): SearchSummary
+
+    companion object {
+        fun Ref.toDocument(repositoryDocument: RepositoryDocument) =
+            RefDocument(repositoryDocument, type, name, commitId, link)
+
+        fun RefDocument.toDto() = when (type) {
+            RefType.BRANCH -> Branch(name, hash, link, repository.toDto())
+            RefType.TAG -> Tag(name, hash, link, repository.toDto())
+        }
+
+        fun Commit.toDocument(repositoryDocument: RepositoryDocument) = CommitDocument(
+            repositoryDocument, id, message, date, author.toDocument(), parents, link
+        )
+
+        fun CommitDocument.toDto() =
+            Commit(hash, message, date, author.toDto(), parents, link, repository.toDto())
+
+        fun PullRequest.toDocument(repositoryDocument: RepositoryDocument) = PullRequestDocument(
+            repositoryDocument,
+            index,
+            title,
+            description,
+            author.toDocument(),
+            source,
+            target,
+            assignees.map { it.toDocument() },
+            reviewers.map { it.toDocument() },
+            status,
+            createdAt,
+            updatedAt,
+            link
+        )
+
+        fun PullRequestDocument.toDto() = PullRequest(
+            index,
+            title,
+            description,
+            author.toDto(),
+            source,
+            target,
+            assignees.map { it.toDto() },
+            reviewers.map { it.toDto() },
+            status,
+            createdAt,
+            updatedAt,
+            link,
+            repository.toDto()
+        )
+
+        private fun PullRequestReviewer.toDocument() = PullRequestReviewerDocument(user.toDocument(), approved)
+
+        private fun PullRequestReviewerDocument.toDto() = PullRequestReviewer(user.toDto(), approved)
+
+        private fun User.toDocument() = UserDocument(name, avatar)
+
+        private fun UserDocument.toDto() = User(name, avatar)
+
+        fun RepositoryDocument.toDto() = Repository(sshUrl, link, avatar)
+    }
 }
